AdminFilter: Skip URI scheme during redirect prefix detection.
This fixes the case where a redirect to, say, /index.xhtml comes out
as http:/index.xhtml instead of the correct URI.  This usually happens
when the application prefix is empty and its path is /, which matches
the first slash after the URI scheme.  This patch handles that case by
skipping over the URI scheme when looking for the start of the prefix
in the request URI.

If the scheme cannot be detected in the request URI, the behavior is
unchanged.

Fixes #160.